RatePlans and Tiers

+++ src/main/java/odm/voltaire/models/SuspensionCredit.java
@@ -1,5 +1,6 @@
 package odm.voltaire.models;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.AllArgsConstructor;
@@ -10,5 +11,5 @@ import lombok.Data;
 public class SuspensionCredit {
   LocalDate startDate;
   LocalDate endDate;
-  double amount;
+  BigDecimal amount;
 }
